Reorganize controllers
