+++ server/src/controllers/app/index.js
@@ -1,0 +1,5 @@
+const users = require('./users');
+
+module.exports = {
+  users
+}
