+++ server/src/controllers/cronofy/index.js
@@ -1,0 +1,9 @@
+const auth = require('./auth');
+const events = require('./events');
+const users = require('./users');
+
+module.exports = {
+  auth,
+  events,
+  users
+}
